Ensure that scr/version is up-to-date when building
This is an experiment to see if it forces me to actually have updated
the version before I build. The idea is that I need to actually commit
the version which will make it more likely that all instances can pull
down the code and display the correct version if I've done it myself.

It uses `git show` to grab the committed contents of src/version then
checks to see if it matches today's date.